Restructure tests for better organization and maintainability
