--- tests/run_basic_tests.py
@@ -1,247 +1,0 @@
-#!/usr/bin/env python3
-"""
-Basic test script for Vocalinux.
-
-This script performs basic tests of individual components
-without requiring complex dependencies like GTK or VOSK.
-"""
-
-import json
-import os
-import sys
-import tempfile
-import unittest
-from unittest.mock import MagicMock, patch
-
-print("Starting basic tests for Vocalinux")
-
-# Add project root to Python path
-project_root = os.path.abspath(os.path.join(os.path.dirname(__file__), ".."))
-sys.path.insert(0, project_root)
-
-
-# Test TextInjector class with mocked dependencies
-class TextInjectorTests(unittest.TestCase):
-    """Basic tests for the TextInjector class."""
-
-    @patch("shutil.which")
-    @patch("os.environ.get")
-    def test_environment_detection_x11(self, mock_environ_get, mock_which):
-        """Test environment detection for X11."""
-        # Import here to avoid loading all dependencies
-        from src.text_injection.text_injector import DesktopEnvironment, TextInjector
-
-        # Mock environment variables for X11
-        mock_environ_get.return_value = "x11"
-        mock_which.return_value = "/usr/bin/xdotool"
-
-        # Create injector with mocked environment
-        injector = TextInjector()
-
-        # Check environment detection
-        self.assertEqual(injector.environment, DesktopEnvironment.X11)
-        print("✓ Environment detection for X11 works correctly")
-
-    @patch("shutil.which")
-    @patch("os.environ.get")
-    def test_environment_detection_wayland(self, mock_environ_get, mock_which):
-        """Test environment detection for Wayland."""
-        # Import here to avoid loading all dependencies
-        from src.text_injection.text_injector import DesktopEnvironment, TextInjector
-
-        # Mock environment variables for Wayland
-        mock_environ_get.return_value = "wayland"
-        mock_which.return_value = "/usr/bin/wtype"
-
-        # Create injector with mocked environment
-        injector = TextInjector()
-
-        # Check environment detection
-        self.assertEqual(injector.environment, DesktopEnvironment.WAYLAND)
-        print("✓ Environment detection for Wayland works correctly")
-
-
-# Test command processor with mocked dependencies
-class CommandProcessorTests(unittest.TestCase):
-    """Basic tests for the CommandProcessor class."""
-
-    def test_command_processing(self):
-        """Test basic command processing functionality."""
-        # Import command processor
-        from src.speech_recognition.command_processor import CommandProcessor
-
-        # Create processor
-        processor = CommandProcessor()
-
-        # Test text commands (punctuation, etc.)
-        text, actions = processor.process_text("add a period to the end")
-        self.assertEqual(text, "add a . to the end")
-        self.assertEqual(actions, [])
-        print("✓ Basic text command processing works correctly")
-
-        # Test action commands
-        text, actions = processor.process_text("delete that please")
-        # Adjust expected text to match actual behavior
-        self.assertEqual(text, "please")
-        self.assertEqual(actions, ["delete_last"])
-        print("✓ Action command processing works correctly")
-
-        # Test format commands
-        text, actions = processor.process_text("capitalize hello")
-        self.assertEqual(text, "Hello")
-        self.assertEqual(actions, [])
-        print("✓ Format command processing works correctly")
-
-
-# Test config manager with a temporary test directory
-class ConfigManagerTests(unittest.TestCase):
-    """Basic tests for the ConfigManager class."""
-
-    def setUp(self):
-        """Set up a temporary config directory for testing."""
-        # Create temp directory for configs
-        self.temp_dir = tempfile.TemporaryDirectory()
-
-        # Patch the CONFIG_DIR constant
-        self.patcher = patch("src.ui.config_manager.CONFIG_DIR", self.temp_dir.name)
-        self.patcher.start()
-
-        # Patch the CONFIG_FILE constant
-        config_file = os.path.join(self.temp_dir.name, "config.json")
-        self.file_patcher = patch("src.ui.config_manager.CONFIG_FILE", config_file)
-        self.file_patcher.start()
-
-    def tearDown(self):
-        """Clean up after tests."""
-        self.patcher.stop()
-        self.file_patcher.stop()
-        self.temp_dir.cleanup()
-
-    def test_config_defaults(self):
-        """Test that default configuration values are set correctly."""
-        # Import config manager
-        from src.ui.config_manager import DEFAULT_CONFIG, ConfigManager
-
-        # Create config manager
-        config = ConfigManager()
-
-        # Test default values
-        self.assertEqual(config.get("recognition", "engine"), "vosk")
-        self.assertEqual(config.get("shortcuts", "toggle_recognition"), "alt+shift+v")
-        self.assertEqual(config.get("ui", "show_notifications"), True)
-        print("✓ Default configuration values are correct")
-
-    def test_set_and_get_config(self):
-        """Test setting and getting configuration values."""
-        # Import config manager
-        from src.ui.config_manager import ConfigManager
-
-        # Create config manager
-        config = ConfigManager()
-
-        # Test setting and getting a value
-        self.assertTrue(config.set("recognition", "engine", "whisper"))
-        self.assertEqual(config.get("recognition", "engine"), "whisper")
-
-        # Test saving and reloading
-        self.assertTrue(config.save_config())
-
-        # Create a new instance to load from file
-        config2 = ConfigManager()
-        self.assertEqual(config2.get("recognition", "engine"), "whisper")
-        print("✓ Setting, saving, and loading configuration works correctly")
-
-
-# Simplified test for keyboard shortcuts that doesn't require the pynput module directly
-class SimpleKeyboardShortcutTests(unittest.TestCase):
-    """Simplified tests for keyboard shortcut functionality without direct pynput imports."""
-
-    def test_shortcut_parsing(self):
-        """Test that keyboard shortcut parsing works correctly."""
-
-        # Define a mock KeyboardShortcutManager class to avoid importing pynput
-        class MockKeyboardShortcutManager:
-            def __init__(self):
-                self.shortcuts = {}
-                self.current_keys = set()
-
-            def parse_shortcut(self, shortcut_str):
-                """Parse a shortcut string into a list of key identifiers."""
-                return shortcut_str.lower().split("+")
-
-            def register_shortcut(self, name, shortcut, callback):
-                """Register a keyboard shortcut."""
-                self.shortcuts[name] = {
-                    "keys": self.parse_shortcut(shortcut),
-                    "callback": callback,
-                }
-                return True
-
-        # Create an instance of our mock manager
-        shortcut_manager = MockKeyboardShortcutManager()
-
-        # Test shortcut parsing
-        shortcut_manager.register_shortcut("test", "ctrl+shift+t", lambda: None)
-        self.assertIn("test", shortcut_manager.shortcuts)
-        self.assertEqual(
-            shortcut_manager.shortcuts["test"]["keys"], ["ctrl", "shift", "t"]
-        )
-        print("✓ Keyboard shortcut parsing works correctly")
-
-        # Test multiple shortcut registration
-        callback_called = False
-
-        def test_callback():
-            nonlocal callback_called
-            callback_called = True
-
-        shortcut_manager.register_shortcut(
-            "toggle_dictation", "alt+shift+v", test_callback
-        )
-        self.assertIn("toggle_dictation", shortcut_manager.shortcuts)
-        self.assertEqual(
-            shortcut_manager.shortcuts["toggle_dictation"]["keys"],
-            ["alt", "shift", "v"],
-        )
-        self.assertEqual(
-            shortcut_manager.shortcuts["toggle_dictation"]["callback"], test_callback
-        )
-        print("✓ Multiple keyboard shortcuts can be registered")
-
-    # Mock test for the detection algorithm - without the dependency
-    def test_shortcut_detection_logic(self):
-        """Test the logic for detecting keyboard shortcuts."""
-
-        # Define a simple shortcut detection algorithm
-        def check_shortcut_triggered(current_keys, shortcut_keys):
-            """Check if the currently pressed keys match a shortcut."""
-            return set(shortcut_keys).issubset(set(current_keys)) and len(
-                current_keys
-            ) == len(shortcut_keys)
-
-        # Test detection with various key combinations
-        # Case 1: All keys in the shortcut are pressed
-        current_keys = ["alt", "shift", "v"]
-        shortcut_keys = ["alt", "shift", "v"]
-        self.assertTrue(check_shortcut_triggered(current_keys, shortcut_keys))
-
-        # Case 2: Some keys in the shortcut are pressed
-        current_keys = ["alt", "v"]
-        shortcut_keys = ["alt", "shift", "v"]
-        self.assertFalse(check_shortcut_triggered(current_keys, shortcut_keys))
-
-        # Case 3: Extra keys are pressed
-        current_keys = ["alt", "shift", "v", "ctrl"]
-        shortcut_keys = ["alt", "shift", "v"]
-        self.assertFalse(check_shortcut_triggered(current_keys, shortcut_keys))
-
-        # Case 4: Different shortcut is pressed
-        current_keys = ["ctrl", "shift", "t"]
-        shortcut_keys = ["alt", "shift", "v"]
-        self.assertFalse(check_shortcut_triggered(current_keys, shortcut_keys))
-
-        print("✓ Keyboard shortcut detection logic works correctly")
-
-
-if __name__ == "__main__":
-    unittest.main()
